==> notebooks/4.1_visualizations_notebook.Rmd <==
---
title: "4.1 Visualizations"
output: html_notebook
---

# Load Environment
```{r message=FALSE, warning=FALSE, results='hide'}
# loading environment 
setwd("~/Desktop/Code/Thesis")
load("~/Desktop/Code/Thesis/Data_for_Analysis/bfinal.Rdata")
load("~/Desktop/Code/Thesis/Data_for_Analysis/eg.rda")
load("~/Desktop/Code/Thesis/Data_for_Analysis/pdau.Rdata")
load("~/Desktop/Code/Thesis/Data_for_Analysis/blong.Rdata")

library(tidyverse)
library(sf)
library(tmap)
library(rgdal)
library(gifski)
library(broom)
library(janitor)
library(knitr)
```

## Load spatial objects
```{r message=FALSE, warning=FALSE, results='hide'}
# get_X_Y_coordinates <- function(x) {
#   sftype <- as.character(sf::st_geometry_type(x, by_geometry = FALSE))
#   if(sftype == "POINT") {
#     xy <- as.data.frame(sf::st_coordinates(x))
#     dplyr::bind_cols(x, xy)
#   } else {
#     x
#   }
# }
# 
# sf_fisbroker <- function(url) {
#   typenames <- basename(url)
#   url <- httr::parse_url(url)
#   url$query <- list(service = "wfs",
#                     version = "2.0.0",
#                     request = "GetFeature",
#                     srsName = "EPSG:25833",
#                     TYPENAMES = typenames)
#   request <- httr::build_url(url)
#   print(request)
#   out <- sf::read_sf(request)
#   out <- sf::st_transform(out, 3035)
#   out <- get_X_Y_coordinates(out)
#   out <- st_as_sf(as.data.frame(out))
#   return(out)
# }
# 
# lor1 <- sf_fisbroker("https://fbinter.stadt-berlin.de/fb/wfs/data/senstadt/s_lor_plan")
# save(lor1, file="~/Desktop/Code/Thesis/Data_for_Analysis/lor1.Rdata")
load("~/Desktop/Code/Thesis/Data_for_Analysis/lor1.Rdata")

lor1 <- lor1 %>% 
  separate(gml_id, c("s_lor_plan", "gml_id"), sep=11) %>% 
  st_drop_geometry() %>% 
  as.data.frame()
# file previously downloaded with sf_fisbroker (see script 1.1)
lor <- st_as_sf(readOGR("~/Desktop/Code/Thesis/shapefiles/lor.shp")) %>% rename(RAUMID = gml_id)
lor <- left_join(as.data.frame(lor), lor1[c(2:6)], by=c("RAUMID" = "gml_id"))
lor$RAUMID <- as.numeric(lor$RAUMID)

# joining objects to shapefile
b_sf <- left_join(bfinal, lor, by=c("RAUMID")) %>% st_as_sf()
bfinal[, 8:16][is.na(bfinal[, 8:16])] <- 0
bwhole <- st_cast(lor %>% st_as_sf, to="POLYGON")

bL_sf <- left_join(dfL_full, lor, by=c("RAUMID")) %>% st_as_sf()
bL_sf$gstatus <- as.factor(bL_sf$gstatus)

stagenames = c("Stage 0",
               "Stage 1 — Pre-gentrification ",
               "Stage 2 — Pioneers Invade",
               "Stage 3 — Pioneers dominate, Initial residents displaced", 
               "Stage 4 — Gentrifiers invade",
               "Stage 5 — Gentrifiers dominate, pioneers displaced")
```
```{r}
ggplot() + geom_sf(data=b_sf, aes(fill=gcode %>% as.factor()), size=.1) + 
  scale_fill_brewer(palette = "OrRd", name="Gentrification Status", labels=stagenames)
```



# facet plot of gentrification status
```{r}
gg <- ggplot() + 
  geom_sf(data=bL_sf, aes(fill=gstatus), size=.1) + 
  facet_wrap(~Year, nrow=4) + 
  scale_fill_brewer(palette = "OrRd", name="Gentrification Status", labels=stagenames) + 
  theme(legend.position = c(0.513, 0.15), 
        legend.title=element_text(size=33), 
        legend.text=element_text(size=27), 
        strip.text = element_text(size=38)) + 
  theme(axis.text.x = element_blank(), axis.text.y = element_blank()) + 
  theme(legend.key.size = unit(3, "cm"))

ggsave(filename="~/Desktop/dg_map.png", gg, width=40, height=48, limitsize = FALSE)
```

# exploratory analysis 

## isolating areas that have gcode 0 in 2019
```{r}
eg_lor_full <- eg_lor_full %>% rename(RAUMID = ID)
eg_lor_full$RAUMID <- as.numeric(eg_lor_full$RAUMID)
highval <- b_sf %>% filter(b_sf$gcode == 0) %>% left_join(., eg_lor_full, by=c("RAUMID")) %>% dplyr::select(c(1, 8, 30:31, 43, 45))
```

## looking at eg status of areas by year
```{r}
y <- bL_sf %>% 
  select(2, 8:9, 12) %>% 
  split(bL_sf$Year) %>%
  map(filter, is_eg_byyear == 1) %>% 
  map(tabyl, gstatus) 

seq <- seq(2, 20, 2)
seq_1 <- seq(3, 21, 2)
y_df <- y[[1]] %>% rename("2001"=2, "2001 %" = 3) %>% 
  left_join(., y[[2]], by=c("gstatusgeometry")) %>% rename("2003"=4, "2003 %" = 5) %>% 
  left_join(., y[[3]], by=c("gstatusgeometry")) %>% rename("2005"=6, "2005 %" = 7) %>% 
  left_join(., y[[4]], by=c("gstatusgeometry")) %>% rename("2007"=8, "2007 %" = 9) %>% 
  left_join(., y[[5]], by=c("gstatusgeometry")) %>% rename("2009"=10, "2009 %" = 11) %>% 
  left_join(., y[[6]], by=c("gstatusgeometry")) %>% rename("2011"=12, "2011 %" = 13) %>% 
  left_join(., y[[7]], by=c("gstatusgeometry")) %>% rename("2013"=14, "2013 %" = 15) %>% 
  left_join(., y[[8]], by=c("gstatusgeometry")) %>% rename("2015"=16, "2015 %" = 17) %>% 
  left_join(., y[[9]], by=c("gstatusgeometry")) %>% rename("2017"=18, "2017 %" = 19) %>% 
  left_join(., y[[10]], by=c("gstatusgeometry")) %>% rename("2019"=20, "2019 %" = 21) %>% 
  select(1, seq, seq_1)
y_df <- cbind(t(y_df %>% select(1:11)) %>% as.data.frame(), t(y_df %>% select(1, 12:21)) %>% as.data.frame())
colnames(y_df) <- y_df[1,]
y_df <- y_df[-c(1),]
colnames(y_df)[1:6] <- paste(c("Number in Stage "), colnames(y_df)[1:6])
colnames(y_df)[7:12] <- paste(c("Percent in Stage "), colnames(y_df)[7:12])
y_df[,c(7:12)] <- (sapply(y_df[,c(7:12)],as.numeric) * 100) %>% 
  round(digits=1)

y_df <- y_df %>% select(1, 7, 2, 8, 3, 9, 4, 10, 5, 11, 6, 12)

evens <- seq(2, 12, 2)
test <- y_df

write.table((y_df), file = "~/Desktop/Code/Thesis/misc_plots/y_df.txt", sep = ",", quote = FALSE, row.names = F)
```


```{r}

```
